:hammer: Add a stream method to the Articles class
Because I might want to be loading up a *lot* of articles, and might want to
do it slowly in the background of some other process, let's throw in a way
of getting articles a page at a time, rather than as one honking great list.

This will save on memory and makes it easier to pull in articles and stream
them out to a database (for example).